move files in common
