+++ packages/twenty-server/src/engine/core-modules/calendar/timeline-calendar-event.module.ts
@@ -4,8 +4,8 @@ import { UserModule } from 'src/engine/core-modules/user/user.module';
 import { TimelineCalendarEventResolver } from 'src/engine/core-modules/calendar/timeline-calendar-event.resolver';
 import { TimelineCalendarEventService } from 'src/engine/core-modules/calendar/timeline-calendar-event.service';
 import { TwentyORMModule } from 'src/engine/twenty-orm/twenty-orm.module';
-import { CalendarEventWorkspaceEntity } from 'src/modules/calendar/standard-objects/calendar-event.workspace-entity';
 import { PersonWorkspaceEntity } from 'src/modules/person/standard-objects/person.workspace-entity';
+import { CalendarEventWorkspaceEntity } from 'src/modules/calendar/common/standard-objects/calendar-event.workspace-entity';
 
 @Module({
   imports: [
