+++ packages/twenty-server/src/modules/calendar/calendar-event-cleaner/calendar-event-cleaner.module.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common';
 import { TwentyORMModule } from 'src/engine/twenty-orm/twenty-orm.module';
 import { DeleteConnectedAccountAssociatedCalendarDataJob } from 'src/modules/calendar/calendar-event-cleaner/jobs/delete-connected-account-associated-calendar-data.job';
 import { CalendarEventCleanerService } from 'src/modules/calendar/calendar-event-cleaner/services/calendar-event-cleaner.service';
-import { CalendarEventWorkspaceEntity } from 'src/modules/calendar/standard-objects/calendar-event.workspace-entity';
+import { CalendarEventWorkspaceEntity } from 'src/modules/calendar/common/standard-objects/calendar-event.workspace-entity';
 
 @Module({
   imports: [TwentyORMModule.forFeature([CalendarEventWorkspaceEntity])],
