+++ packages/twenty-server/src/modules/calendar/utils/format-google-calendar-event.util.ts
@@ -1,8 +1,8 @@
 import { calendar_v3 as calendarV3 } from 'googleapis';
 import { v4 } from 'uuid';
 
-import { CalendarEventParticipantResponseStatus } from 'src/modules/calendar/standard-objects/calendar-event-participant.workspace-entity';
 import { CalendarEventWithParticipants } from 'src/modules/calendar/types/calendar-event';
+import { CalendarEventParticipantResponseStatus } from 'src/modules/calendar/common/standard-objects/calendar-event-participant.workspace-entity';
 
 export const formatGoogleCalendarEvent = (
   event: calendarV3.Schema$Event,
